server/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
motor==3.1.2
pymongo==4.3.3
sqlalchemy==2.0.9
python-dotenv==1.0.0
httpx==0.24.0
openai==0.27.4
PyJWT==2.6.0
